# openapi-schema.yaml
openapi: 3.0.3
info:
  title: Training Manager API
  version: 1.0.0
  description: API pour la gestion d entrainements
paths:
...
  /api/v1/events/{id}/generate-training/:
    post:
      operationId: events_generate_training_create
      description: Generate detailed Rounds and Exercises with AI for an Event. Optionally
        accepts an `additional_prompt` (max 2000 chars) appended to the LLM user prompt
        after the structured context.
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this event.
        required: true
      tags:
      - events
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateTrainingRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/GenerateTrainingRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/GenerateTrainingRequest'
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenerateTrainingResponse'
          description: Training session generated successfully
        '400':
          description: Invalid request body
        '403':
          description: Not a manager of this event team
        '409':
          description: Event already has rounds
        '500':
          description: AI configuration error
        '502':
          description: AI service error
  /api/v1/events/{id}/rounds/reorder/:
    post:
      operationId: events_rounds_reorder_create
      description: Atomically reorder the Rounds attached to this Event. `round_ids`
        must contain exactly the IDs of the Rounds currently attached, in the desired
        final order. Round.order is set to 1..N matching list position, in a single
        transaction.
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this event.
        required: true
      tags:
      - events
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReorderRoundsRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/ReorderRoundsRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ReorderRoundsRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '204':
          description: Rounds reordered
        '400':
          description: 'Validation error. `body.code` is one of: `empty_list`, `duplicate_id`,
            `scope_mismatch`, `incomplete_reorder`.'
        '403':
          description: Not a manager of this event team
  /api/v1/exercises/:
    get:
      operationId: exercises_list
      description: CRUD complet pour Exercise.
      parameters:
      - in: query
        name: energysegment
        schema:
          type: integer
      - in: query
        name: language
        schema:
          type: string
          enum:
          - en
          - es
          - fr
          - it
          - nl
        description: |-
          * `fr` - Français
          * `nl` - Nederlands
          * `en` - English
          * `it` - Italiano
          * `es` - Español
      - in: query
        name: modality
        schema:
          type: integer
      - name: ordering
        required: false
        in: query
        description: Which field to use when ordering the results.
        schema:
          type: string
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      - name: page_size
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      - name: search
        required: false
        in: query
        description: A search term.
        schema:
          type: string
      tags:
      - exercises
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedExerciseList'
          description: ''
    post:
      operationId: exercises_create
      description: CRUD complet pour Exercise.
      tags:
      - exercises
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Exercise'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Exercise'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Exercise'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exercise'
          description: ''
...
components:
  schemas:
...
    EnergySegment:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        abv:
          type: string
          maxLength: 10
        description:
          type: string
          nullable: true
          maxLength: 200
        energy_system:
          allOf:
          - $ref: '#/components/schemas/EnergySystem'
          readOnly: true
        energy_system_id:
          type: integer
          writeOnly: true
          nullable: true
        is_active:
          type: boolean
          readOnly: true
      required:
      - abv
      - energy_system
      - id
      - is_active
...
    EnergySystem:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          readOnly: true
        is_active:
          type: boolean
          readOnly: true
      required:
      - id
      - is_active
      - name
...
    Exercise:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        order:
          type: integer
          maximum: 9223372036854775807
          minimum: -9223372036854775808
          format: int64
        repetition:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
        distance:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
        notes:
          type: string
          maxLength: 200
        t_start:
          type: string
          nullable: true
          description: MM:SS format, e.g. 1:30.
          pattern: ^\d{1,3}:[0-5]\d$
        t_break:
          type: string
          nullable: true
          description: MM:SS format, e.g. 1:30.
          pattern: ^\d{1,3}:[0-5]\d$
        modality:
          allOf:
          - $ref: '#/components/schemas/Modality'
          readOnly: true
        modality_id:
          type: integer
          writeOnly: true
          nullable: true
        energysegment:
          allOf:
          - $ref: '#/components/schemas/EnergySegment'
          readOnly: true
        energysegment_id:
          type: integer
          writeOnly: true
          nullable: true
        round_id:
          type: integer
          writeOnly: true
          description: Optional. If provided on POST, the newly created Exercise is
            atomically attached to this Round. The request user must manage at least
            one team of at least one Event linked to this Round (library rounds with
            no events are accepted as-is). Ignored on PATCH/PUT.
        language:
          $ref: '#/components/schemas/LanguageEnum'
        usage_count:
          type: integer
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - created_at
      - energysegment
      - id
      - language
      - modality
      - updated_at
      - usage_count
...
    GenerateTrainingRequest:
      type: object
      description: |-
        Optional payload for POST /events/{id}/generate-training/.

        `additional_prompt` is appended to the LLM user prompt after the
        structured context (sport, level, duration, catalogs). Empty/missing
        is a no-op so older clients without a body remain compatible.
      properties:
        additional_prompt:
          type: string
          default: ''
    GenerateTrainingResponse:
      type: object
      properties:
        rounds_created:
          type: integer
        exercises_created:
          type: integer
        exercises_reused:
          type: integer
        rationale:
          type: string
        model:
          type: string
        tokens_used:
          $ref: '#/components/schemas/GenerateTrainingTokensUsed'
      required:
      - exercises_created
      - exercises_reused
      - model
      - rationale
      - rounds_created
      - tokens_used
    GenerateTrainingTokensUsed:
      type: object
      properties:
        input:
          type: integer
        output:
          type: integer
      required:
      - input
      - output
...
    LanguageEnum:
      enum:
      - fr
      - nl
      - en
      - it
      - es
      type: string
      description: |-
        * `fr` - Français
        * `nl` - Nederlands
        * `en` - English
        * `it` - Italiano
        * `es` - Español
...
    Modality:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          readOnly: true
        sport:
          allOf:
          - $ref: '#/components/schemas/Sport'
          readOnly: true
        is_active:
          type: boolean
          readOnly: true
      required:
      - id
      - is_active
      - name
      - sport
...
    PaginatedExerciseList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/Exercise'
...
    ReorderRoundsRequest:
      type: object
      description: |-
        Body for POST /events/{id}/rounds/reorder/.

        `round_ids` must contain exactly the IDs of the Rounds attached to the
        Event in the desired final order. The view enforces scope/completeness
        in addition to this serializer's syntactic validation.
      properties:
        round_ids:
          type: array
          items:
            type: integer
            minimum: 1
      required:
      - round_ids
...
    Sport:
      type: object
      description: |-
        Public read serializer: 'name' is auto-localised by modeltranslation.
        energy_systems exposed as PK list (frontend joins via /energy-systems/).
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          readOnly: true
        slug:
          type: string
          readOnly: true
          pattern: ^[-a-zA-Z0-9_]+$
        is_active:
          type: boolean
          readOnly: true
        energy_systems:
          type: array
          items:
            type: integer
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - created_at
      - energy_systems
      - id
      - is_active
      - name
      - slug
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT